For #1694, Refine API for nb_bytes

+++ trunk/src/kernel/srs_kernel_buffer.hpp
@@ -41,8 +41,7 @@ public:
     /**
      * get the number of bytes to code to.
      */
-    // TODO: FIXME: change to uint64_t.
-    virtual int nb_bytes() = 0;
+    virtual uint64_t nb_bytes() = 0;
     /**
      * encode object to bytes in SrsBuffer.
      */
